Spring Security Login y Register listo

+++ src/main/java/com/buildoc/buildocDemo/services/RolServices.java
@@ -11,7 +11,7 @@ public interface RolServices {
 
     public void crearRol(Rol rol);
 
-    public Rol obtenerRolPorId(Long id);
+    public Rol obtenerRolPorId(Integer id);
 
     public void actualizarRol(Rol rol);
